refactor(agent): split stop and status into their own packages
Move the 'agent stop' and 'agent status' client commands into
pkg/controller/agent/stop and pkg/controller/agent/status, following the
unlock split. Both are pure clients (only agentapi), so they need no shared
helpers.

reset now calls stop.New().Run instead of the in-package Stop method. The
serve-over-socket status tests stay in core (they need the unexported
listen/serve) and now drive the agent with agentapi.Send directly.
