+++ pkg/controller/agent/stop/stop_internal_test.go
@@ -1,4 +1,4 @@
-package agent
+package stop
 
 import (
 	"log/slog"
@@ -10,7 +10,7 @@ import (
 // (returns nil) rather than failing, mirroring 'systemctl stop'.
 func TestStop_notRunning(t *testing.T) {
 	t.Setenv("GHTKN_AGENT_SOCKET", filepath.Join(t.TempDir(), "absent.sock"))
-	if err := New().Stop(t.Context(), slog.New(slog.DiscardHandler)); err != nil {
-		t.Fatalf("Stop with no agent running must return nil, got: %v", err)
+	if err := New().Run(t.Context(), slog.New(slog.DiscardHandler)); err != nil {
+		t.Fatalf("Run with no agent running must return nil, got: %v", err)
 	}
 }
